Add --index-url option for install
Note (presently) am not supporting this option on the inject command. If
you inject a package into an app then the inject will use the same index
as used for the original install.

To support on the inject is awkward because we would have to track the
index-url per package, including those that change version with an index
change.

Addresses issue #1.