Remove --pyenv options
Am removing pyenv options because you can just specify the full path to
the pyenv python version anyhow so no real loss of functionality (other
than you need to specify an explicit path and it is a longer argument to
type) and there are other equivalent python installers such as hatch and
rye (and soon to be uv) so best to not give special privilege to just
pyenv.

So, e.g, instead of "--pyenv=3.12" you now need to specify
"--python=~/.pyenv/versions/3.12/bin/python".