Fix creation on temporary lock file
A lock file is used to ensure we don't clash when creating venvs
possibly in parallel when the user is running pipux, e.g. in parallel in
separate windows. This lock file was being created in a personal user
directory but really it should be created for the pipxu home/base
directory, regardless of any particular user. E.g. if two users are
pointing to the same home/base dir then they should be locked out also.

So this change relocates the lock file to the home dir instead.

Fixes issue #8.